Upgraded exception types
Added templating for specific exception type data payloads, streamlined
the generation of error messages for logging, and added an overload for
std::exception::what that allows a quick dump of the message when an
exception isn't handled. The mutable version of what is not an overload
so must be printed manually in a catch block.