Added BandedMatrix operators
Added some useful arithmetic operations to BandedMatrix that don't
require knowledge about the bandedness of the storage. The non-member
operators are in-class friends due to C++ overload resolution rules as
for DenseMatrix, if passing mixed BandedMatrix types the deduction would
fail and there would be no legal candidate for the operator.