Randomiser for containers, matrix-data setting
Added operation to supply underlying data directly to matrices (if the
new data matches the size of the existing underlying data). Enabled
randomisation routine for Vector, and by extension to the matrix types.
Iterator access enabled to BandedMatrix for consistency with
DenseMatrix, though it's not clear how useful that is at this stage.