Separated test result generation and analysis
TestResult now creates a tree structure for a given test suite (to be
extended to completely generic heirarchies of suites), and the reporting
is left to a separate class. The TestReporter can trace each test and
identify which report results/skips, and summarise fully qualified names
of tests which failed or were skipped.