BandedMatrix element access adjusted
Some ambiguity existed in which operator[] would be called, so it's
better to have an explicitly-marked unsafe getter if anyone chooses to
use that, with an explicit setter that returns a flag for success and
never throws. The operator[] always returns a temporary and only throws
if the overall matrix bounds are violated.